+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Celestial GTK Theme will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Celestial window decorations (Aurorae) for KDE Plasma, applied by the global themes
+
 ## [1.4.0] - 2026-07-21
 
 ### Added